[GEOT-6925] Factor out envelope transformation, from CRS to its own class, split blocks of transform method out to dedicated methods
